+++ Tools/calibrator.py
@@ -2,10 +2,8 @@ import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
 import scipy as sc
-from GetData import cutData
+from Tools.GetData import cutData
 from constant import *
-
-
 
 
 '''
@@ -36,12 +34,13 @@ def findPlateau(graph):
     
 
     for x, y in enumerate(graph[:-1], 2):
+
         duration += 1 # Increase counter
         if(x%CALCRATE == 0 ): 
             # If variance has increased beyond our threshold, record identified plateau using the mean
             # Then, reset running values and counter
-            if(not len(results) or variance > MAXVARIANCE and duration > TIMETHRESHOLD):
-                results.append([prevMean, start, x, duration])# Save average as plateau val
+            if(not len(results) or abs(variance) > MAXVARIANCE and duration > TIMETHRESHOLD):
+                results.append([prevMean, start, x, duration]) # Save average as plateau val
 
                 # Reset stats
                 prevMean = -1
@@ -50,7 +49,7 @@ def findPlateau(graph):
                 start = x + CALCTHRESH
 
             
-            if(duration > CALCTHRESH):      # Only calculate at CALCRATE and check if it's been enough time since the last plateau
+            if(duration > CALCTHRESH):      # Check if it's been enough time since the last plateau or if we've reached the end
                 mean = prevMean + ((y - prevMean)/(duration-CALCTHRESH))            # Calculate running mean
                 stdDev = (prevStdDev + (y - prevMean)*(y - mean))                   # Calculate running standard deviation
                 variance = stdDev/(duration-CALCTHRESH)                             # calculate running variance
@@ -61,9 +60,10 @@ def findPlateau(graph):
     
         varianceGraph.append(prevMean)
 
+    results.append([prevMean, start, x, duration])  # Add final level as plateau
     return [results, varianceGraph]
 
-
+#len(graph)-x <= ENDTHRESH
 def filter(input):
     return sc.signal.medfilt(input, FILTWIDTH)
 
@@ -75,9 +75,8 @@ if __name__ == "__main__":
     testfire0 = 'data/calibration-data-raw.csv'
     testfire1 = '2022_2023_Data/Calibration1.csv'
 
-    graph = cutData(pd.read_csv(testfire1,names=['values']))
+    graph = cutData(pd.read_csv(testfire0,names=['values']))
     graph = graph['values'].to_list()
-
 
     res = findPlateau(graph)
