+++ data/processed_data/5-07-22_Test_Fire_Calibration_2_expected_values.csv
@@ -1,0 +1,6 @@
+
+605.46875,,,,605.4688,45.9,,204.1733721
+605.46875,,,,712.89,56.9,,253.1038099
+605.46875,,,,781.25,66.9,,297.5860261
+605.46875,,,,864.26,75.7,,336.7303763
+605.46875,,,,913.09,81.2,,361.1955952
